requirements.txt
stockfish
python-chess
elevenlabs
python-dotenv
opencv-python
roboflow
inference
numpy
langgraph
openai
Flask
